% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwp.R
\name{gwp_value}
\alias{gwp_value}
\title{Pull GWP values}
\usage{
gwp_value(x, type = c("gwp_ar4_100"))
}
\description{
Defaults to AR4 100-year GWP currently.
Accepted values: c("gwp_ar4_100", "gwp_ar4_20", "gwp_ar4_500", "gwp_sar_100")
}
