% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_ghg_names.R
\name{normalize_ghg_names}
\alias{normalize_ghg_names}
\title{Normalize GHG names based on ghg_dictionary}
\usage{
normalize_ghg_names(x)
}
\description{
Normalize GHG names based on ghg_dictionary
}
